+++ CHANGELOG.md
@@ -41,12 +41,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug that created a second result chat message when an actor rolled a critical failure to unstun.
 ### Known Issues
 - Holding still breaks on first round of combat but I have no clue why.
+- The Shape Changer macro throws a lot of errors when merging. Line 350 and 365 where disabled to prevent that until further investigation.
 ### Removed
 - Unshake macros from the compendium were removed and unified into a single macro since the version is now drawn from the settings.
 - Removed v10 update warning.
 ### Deprecated
 - Old macros deprecated. They are still available in a zip file in case s/o wants to check them out.
 - ways to get data from additional stats were completely removed. A migration is provided.
+- Token merging disabled because of the error mentioned above. It'll be reintroduced once the problem was found.
 
 ## [0.18.5] - 2022-08-14
 ☮️ Peace in the world, or the world in pieces. 🕊️